Use project path instead of project name to key sub-projects.
Project path has the format ":subName:subSubName" and ensures two
sub-projects with the same name does not collide.

Keep using project name for the root project (as the path is just ":")

There should be no risk of collision since the root project can not
have ":" in its name, and all sub-project paths will start with ":".